add test for import-export workflow
